Drain keyboard buffer after running program
Sometimes when a program exits stray keystrokes are left in the keyboard
buffer, and the menu then responds to them (e.g. by launching the first
menu item). For example, one of the games crashes (due to unrelated
reasons) and then the first game on the menu automatically launches, and
Blake gets upset. This won't stop the crash but will hopefully stop the
first game from launching when it does.