Hacky workaround for bad monitor overscan
The TV in the family room has a bad case of overscan. Neither the TV
itself, nor the ancient version of Linux (Fedora Core 4) running on the
attached PC seem to let me fix it. As a result, the left hand side of
the menu border is cut off. So the hacky workaround is add a setting
which enables skipping the leftmost/rightmost columns when drawing the
border.